Add files via upload
This is an input file with random initial conditions, which I was using to test the code. 
In the instructions, it said the input file was supposed to be called 'param.input', but I couldn't do this on my computer... so I made it param.txt, no idea what the difference is.